auth_service/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Anis Momin <anis_momin@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
httpx = "^0.27.2"
uvicorn = {extras = ["standard"], version = "^0.30.6"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlmodel = "^0.0.22"
psycopg = {extras = ["binary"], version = "^3.2.1"}
fastapi = "0.111.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
